feat: user, alarms controller에 흩어져있던 auth 통합
